+++ circuits/src/parsing/scanner/static_specs.rs
@@ -1,4 +1,4 @@
-//! A library of parsers, represented as a Hash Map from documented custom
+//! A library of parsers, represented as a Hash Map from documented fixed
 //! tokens (`StdLibParser`) to deterministic minimal automata (`Automata`).
 //! Since these parsers can be relatively costly to generate, the automata are
 //! serialized. The tests check in particular the consistency between the
@@ -21,7 +21,7 @@
 //!    the corresponding serialisation file, and re-run step 5.
 //!
 //! ```shell
-//!    cargo test --lib -p midnight-circuits --release -- --nocapture regex_test automaton_test specs_test
+//!    cargo test --lib -p midnight-circuits --release -- --nocapture regex_test automaton_test static_specs_test
 //! ```
 
 #[cfg(test)]
@@ -32,8 +32,8 @@ use rustc_hash::FxHashMap;
 use super::{
     automaton::Automaton,
     regex::{Regex, RegexInstructions},
+    serialization::Serialize,
 };
-use crate::parsing::scanner::serialization::Serialize;
 
 /// Folder where the serialized automata for the standard library will be
 /// stored.
@@ -110,6 +110,70 @@ pub enum StdLibParser {
     ///   - `"x"` -> 5
     ///   - `"y"` -> 6
     Jwt,
+
+    /// # Description and sources
+    ///
+    /// Format of the Data Group 1 (DG1) of biometric passports, as specified in
+    /// ICAO Doc 9303 for TD3-type documents (machine-readable passports).
+    ///
+    /// The DG1 contains the Machine Readable Zone (MRZ) printed on the
+    /// passport's main page and stored verbatim on the embedded chip. It
+    /// encodes key identity attributes such as the (possibly truncated)
+    /// holder's name, date of birth, nationality, and passport number,
+    /// using a fixed-length ASCII format.
+    ///
+    /// A typical TD3 MRZ looks as follows on passports:
+    ///
+    /// ```text
+    /// PPFRADUPONT<<JEAN<MICHEL<<<<<<<<<<<<<<<<<<<<
+    /// 12AB345678FRA7408122M3101012<<<<<<<<<<<<<<04
+    /// ```
+    ///
+    /// Note that despite being presented as two lines of 44 bytes each, the DG1
+    /// is read and parsed *as a single 88-byte line*.
+    ///
+    /// These 88 bytes, along with the other Data Groups, can be retrieved by
+    /// reading the passport's NFC chip. Integrity and authenticity are ensured
+    /// via the Security Object Document (SOD), which contains the signed
+    /// sequence of hashes of the Data Groups. These signatures can be
+    /// verified using public keys distributed through the ICAO Public Key
+    /// Directory (PKD).
+    ///
+    /// Sources:
+    /// - [ICAO Doc 9303](https://www.icao.int/publications/doc-series/doc-9303)
+    /// - [PKD](https://www.icao.int/icao-pkd)
+    ///
+    /// # Output behaviour
+    ///
+    /// In the MRZ format, only uppercase letters, digits, and `<` (representing
+    /// special characters such as spaces or dashes, as well as padding and
+    /// separators), are used. The following fields are then output as follows;
+    /// note that checksum fields are not mentioned, as they get no output and
+    /// are not verified by the parser.
+    ///  - Passport type (2 bytes; uppercase) -> 1
+    ///  - Issuing country code (3 bytes; uppercase) -> 2
+    ///    + **Note**: this code is ISO 3166-1 alpha-3 compliant.
+    ///  - name field (up to 39 bytes; uppercase and spaces) -> 3 (surname) and
+    ///    4 (given names, if any)
+    ///    + **Note 1**: mononyms, i.e., people having no given names, are
+    ///      ICAO9303 TD3 compliant. In this case, no byte will be output `4`.
+    ///    + **Note 2**: if the credential holder has at least one given name,
+    ///      the credential must include a `<<` separator between the surname
+    ///      and the given names. Names may be truncated if needed to make the
+    ///      separator fit. The whole field is also padded with `<` bytes if no
+    ///      truncation occurred.
+    ///    + **Note 3**: The `<` characters (separator or padding) are not
+    ///      output by this parser. E.g., in `DUPONT<<JEAN<MICHEL`, the 3 `<`
+    ///      get no output. This is because of the impossibility, for a finite
+    ///      automaton to, e.g., decide the third `<` is a padding or, like
+    ///      here, a space before an additional given name.
+    ///  - Passport number (9 bytes; uppercase and digits) -> 5
+    ///  - Nationality (3 bytes; uppercase) -> 6
+    ///  - Date of birth (6 bytes; YYMMDD) -> 7
+    ///  - Sex (1 byte; `M` Male, `F` Female, or `<` Other) -> 8
+    ///  - Date of expiry (6 bytes; YYMMDD) -> 9
+    ///  - Optional data (14 bytes; uppercase, digits, and `<`) -> 10
+    Icao9309Td3Dg1,
 }
 
 #[cfg(test)]
@@ -134,11 +198,18 @@ type ParsingLibrary = FxHashMap<StdLibParser, Automaton>;
 /// components. When serialization is disabled, the automaton will be computed
 /// using the second argument.
 fn spec_library_data() -> LibraryData {
-    &[(
-        StdLibParser::Jwt,
-        &spec_jwt as &'static dyn Fn() -> Regex,
-        include_bytes!("automaton_cache/Jwt") as &'static [u8],
-    )]
+    &[
+        (
+            StdLibParser::Jwt,
+            &spec_jwt,
+            include_bytes!("automaton_cache/Jwt"),
+        ),
+        (
+            StdLibParser::Icao9309Td3Dg1,
+            &spec_icao9303_td3_dg1,
+            include_bytes!("automaton_cache/Icao9309Td3Dg1"),
+        ),
+    ]
 }
 
 /// All automata that can be used as a parsing basis in the standard library.
@@ -157,7 +228,7 @@ pub fn spec_library() -> ParsingLibrary {
             .collect::<FxHashMap<_, _>>()
 }
 
-// Regex formalising the spec of `StdLIbParser::Jwt`.
+/// Regex formalising the spec of `StdLIbParser::Jwt`.
 fn spec_jwt() -> Regex {
     // Content of a basic field (RFC 8259 JSON string), possibly with output if
     // `output` is not 0.
@@ -280,6 +351,79 @@ fn spec_jwt() -> Regex {
     )
 }
 
+/// Regex formalising the spec of `StdLIbParser::ICAO9303DataGroup1`.
+fn spec_icao9303_td3_dg1() -> Regex {
+    // The list of all tolerated passport types. Consulted at this document, Section
+    // 4.4, on Jan. 14, 2026:
+    // https://www.icao.int/sites/default/files/publications/DocSeries/9303_p4_cons_en.pdf
+    // Output 1.
+    let passport_type = Regex::union([
+        "P<".into(), // Legacy denomination of `PP`, before Jan. 2026.
+        "PP".into(), // National/Ordinary passport.
+        "PE".into(), // Emergency passport.
+        "PD".into(), // Diplomatic passport.
+        "PO".into(), // Official/Service passport.
+        "PR".into(), // Refugee passport.
+        "PT".into(), // Alien passport.
+        "PS".into(), // Stateless passport.
+        "PL".into(), // Laissez-passez passport.
+        "PM".into(), // Military passport.
+        "PU".into(), /* Emergency travel document. See: https://www.icao.int/sites/default/files/publications/DocSeries/9303_p8_cons_en.pdf */
+    ]).output(&|_| Some(1));
+    // A non-empty sequence of uppercase letters, with output `output`.
+    let name_block = |output: usize| -> Regex {
+        Regex::uppercase_letter().non_empty_list().output(&|_| Some(output))
+    };
+    // One uppercase letter or a digit, with output `output`.
+    let alphanum = |output: usize| -> Regex {
+        Regex::byte_from((b'A'..=b'Z').chain(b'0'..=b'9')).output(&|_| Some(output))
+    };
+    // A date with the given output, in YYMMDD format.
+    let date = |output: usize| -> Regex { Regex::digit().output(&|_| Some(output)).repeat(6) };
+    // Any passport character.
+    let any = Regex::byte_from((b'A'..=b'Z').chain(b'0'..=b'9').chain(std::iter::once(b'<')));
+
+    // Example to illustrate the code below:
+    // P<FRADUPONT<<JEAN<MICHEL<<<<<<<<<<<<<<<<<<<<
+    // 12AB345678FRA7408122M3101012<<<<<<<<<<<<<<04
+
+    // Mandatory part of the first line of the DG1 (passport type, issuer, surname).
+    // The separators `<` get no output.
+    let line1_prefix = Regex::cat([
+        passport_type,
+        Regex::uppercase_letter().output(&|_| Some(2)).repeat(3),
+        name_block(3).separated_non_empty_list("<".into()),
+    ]);
+    // Given names in the first line of the DG1, prefixed with the separator. The
+    // separators `<` get no output.
+    let given_names = Regex::cat([
+        "<<".into(),
+        name_block(4).separated_non_empty_list("<".into()),
+    ]);
+    // The part of the first line following `line1_prefix`. Considers cases where
+    // given names are present or not.
+    let line1_suffix = given_names.optional().terminated(Regex::word("<").list());
+    // The full first line, with the length constraint.
+    let line1 = line1_prefix.terminated(line1_suffix).and(any.clone().repeat(44));
+
+    // The second line of the DG1. All fields are length constrained.
+    let line2 = Regex::cat([
+        alphanum(5).repeat(9),
+        Regex::digit(), // No-output checksum.
+        Regex::uppercase_letter().output(&|_| Some(6)).repeat(3),
+        date(7),
+        Regex::digit(), // No-output checksum.
+        Regex::byte_from([b'<', b'M', b'F']).output(&|_| Some(8)),
+        date(9),
+        Regex::digit(), // No-output checksum.
+        any.repeat(14).output(&|_| Some(10)),
+        Regex::digit().repeat(2), // No-output checksum.
+    ]);
+
+    // Concatenating the two lines, without a newline character.
+    line1.terminated(line2)
+}
+
 #[cfg(test)]
 /// Re-serialises the data in `checks`, and:
 ///
@@ -302,7 +446,7 @@ fn check_serialization(checks: &ParsingLibrary) {
         let file_name = parser.serialization_file();
         assert!(
                 Path::new(&file_name).exists(),
-                "serialisation file {file_name} does not exist! Follow the documentation of `midnight_circuits::parsing::specs` for instructions on how to add a new parser to the standard library."
+                "serialisation file {file_name} does not exist! Follow the documentation of `midnight_circuits::parsing::scanner::static_specs` for instructions on how to add a new parser to the standard library."
             );
         let previous_data = fs::read(file_name.clone()).unwrap();
         let mut current_data = Vec::new();
@@ -334,11 +478,8 @@ mod tests {
     use rustc_hash::{FxBuildHasher, FxHashMap};
 
     use super::{
-        spec_library_data,
-        StdLibParser::{self, Jwt},
-    };
-    use crate::parsing::scanner::{
-        automaton::Automaton, spec_library, static_specs::check_serialization,
+        super::automaton::Automaton, check_serialization, spec_library, spec_library_data,
+        StdLibParser,
     };
 
     /// Sets up the serialised library (bootstraps it if empty serialisation
@@ -485,15 +626,26 @@ mod tests {
         println!(">> Now configuring the spec library for tests... (using the serialised data)");
         let start = Instant::now();
         let spec_library = spec_library();
-        println!(">> Configuration completed in {:?}", start.elapsed());
+        println!(
+            ">> Configuration completed in {:?}. Automaton breakdown:",
+            start.elapsed()
+        );
+        let mut total = 0;
         for (name, automaton) in &spec_library {
             println!(
-                "  - {:?} automaton: {} states, {} transitions",
+                "  - {:?}: {} states, {} transitions",
                 name,
                 automaton.nb_states,
                 automaton.transitions.values().map(|m| m.len()).sum::<usize>()
-            )
+            );
+            total += automaton.transitions.values().map(|m| m.len()).sum::<usize>()
+                + automaton.final_states.len()
         }
+        println!(
+            ">> Total nb of lookup rows in the chip: {} ≤ 2^{}",
+            total,
+            total.next_power_of_two().trailing_zeros()
+        );
 
         // Tests the `Jwt` spec correctness.
         const FULL_INPUT_JWT: &str = include_str!("specs_examples/jwt/full.txt");
@@ -524,7 +676,95 @@ mod tests {
         ];
         let rejected0: Vec<&str> =
             vec!["hello world", &FULL_INPUT_JWT[..1000], &MINIMAL_JWT[..600]];
+        specs_one_test(&spec_library, StdLibParser::Jwt, &accepted0, &rejected0);
 
-        specs_one_test(&spec_library, Jwt, &accepted0, &rejected0);
+        // Tests the `ICAO9303DataGroup1` spec correctness.
+        let accepted1_raw = include_str!("specs_examples/icao9303_td3_dg1/valid_credentials.txt")
+            .lines()
+            .collect::<Vec<_>>();
+        let accepted1: Vec<(&str, &[(usize, &str)])> = vec![
+            (
+                accepted1_raw[0],
+                &[
+                    (1, "PP"),
+                    (2, "JPN"),
+                    (3, "OKABE"),
+                    (4, "RINTARO"),
+                    (5, "12AB34567"),
+                    (6, "JPN"),
+                    (7, "911214"),
+                    (8, "M"),
+                    (9, "310101"),
+                    (10, "EL<PSY<CONGROO"),
+                ],
+            ),
+            (
+                accepted1_raw[1],
+                &[
+                    (1, "PE"),
+                    (2, "ESP"),
+                    (3, "DELACRUZ"),
+                    (4, "MARIA"),
+                    (5, "UH87G9901"),
+                    (6, "ESP"),
+                    (7, "911214"),
+                    (8, "F"),
+                    (9, "310101"),
+                    (10, "XXV789<<<<<<<<"),
+                ],
+            ),
+            (
+                accepted1_raw[2],
+                &[
+                    (1, "PD"),
+                    (2, "MDG"),
+                    (3, "ANDRIANAMPOINIMERINATOMPOLOINDRINDRA"),
+                    (4, "R"),
+                    (5, "BDL3820HR"),
+                    (6, "FRA"),
+                    (7, "450101"),
+                    (8, "<"),
+                    (9, "600101"),
+                    (10, "<<<<<<<<<<<<<<"),
+                ],
+            ),
+            (
+                accepted1_raw[3],
+                &[
+                    (1, "PO"),
+                    (2, "FRA"),
+                    (3, "NOOOWAYIGOTATRUNCATEDMONONYMRIGH"),
+                    (5, "AAAAAAAAA"),
+                    (6, "FRA"),
+                    (7, "990101"),
+                    (8, "<"),
+                    (9, "300101"),
+                    (10, "<<<<<<<<<<<<<<"),
+                ],
+            ),
+            (
+                accepted1_raw[4],
+                &[
+                    (1, "PR"),
+                    (2, "USA"),
+                    (3, "ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ"),
+                    (5, "PPPPPPPPP"),
+                    (6, "USA"),
+                    (7, "990101"),
+                    (8, "M"),
+                    (9, "300102"),
+                    (10, "<<<<<<<<<<<<<<"),
+                ],
+            ),
+        ];
+        let rejected1 = include_str!("specs_examples/icao9303_td3_dg1/invalid_credentials.txt")
+            .lines()
+            .collect::<Vec<_>>();
+        specs_one_test(
+            &spec_library,
+            StdLibParser::Icao9309Td3Dg1,
+            &accepted1,
+            &rejected1,
+        );
     }
 }